refactor: move loader to server folder
